[nest-analysis] separate temporal analysis. Commented fused-mapping to build dynamically
