+++ include/loop-analysis/bounds-inference.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "loop-analysis/isl-ir.hpp"
+
+LogicalBufOccupancies BoundsInference(LogicalBufOccupancies occupancies);
