+++ include/loop-analysis/fused-mapping-to-ir.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "loop-analysis/isl-ir.hpp"
+
+LogicalBufOccupancies
+OccupanciesFromMapping(const mapping::FusedMapping& mapping,
+                       const WorkloadIR& workload_ir);
