+++ include/loop-analysis/temporal-analysis.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "loop-analysis/isl-ir.hpp"
+
+namespace analysis
+{
+
+std::pair<LogicalBufOccupancies, LogicalBufFills>
+TemporalReuseAnalysis(const LogicalBufOccupancies& occupancies);
+
+} // namespace analysis
