+++ src/SConscript
@@ -138,6 +138,8 @@ loop-analysis/point-set-multi-aahr.cpp
 loop-analysis/nest-analysis-tile-info.cpp
 loop-analysis/nest-analysis.cpp
 loop-analysis/spatial-analysis.cpp
+loop-analysis/temporal-analysis.cpp
+loop-analysis/workload-ir.cpp
 sparse-analysis/state.cpp
 sparse-analysis/sparse-analysis.cpp
 sparse-analysis/compute-gs-analyzer.cpp
