+++ src/loop-analysis/bounds-inference.cpp
@@ -1,0 +1,6 @@
+#include "loop-analysis/isl-ir.hpp"
+
+LogicalBufOccupancies BoundsInference(LogicalBufOccupancies occupancies)
+{
+
+}
