+++ src/loop-analysis/nest-analysis.cpp
@@ -52,6 +52,7 @@
 #include "loop-analysis/isl-ir.hpp"
 #include "loop-analysis/nest-analysis.hpp"
 #include "loop-analysis/spatial-analysis.hpp"
+#include "loop-analysis/temporal-analysis.hpp"
 #include "mapping/fused-mapping.hpp"
 
 bool gTerminateEval = false;
@@ -90,105 +91,12 @@ bool gEnableImperfectCycleCount = false;
 namespace analysis
 {
 
-std::pair<Occupancy, Fill> FillFromOccupancy(Occupancy);
-
-LogicalBufOccupancies
-RemoveIneffectualTemporalDims(const LogicalBufOccupancies& occupancies);
-
-std::pair<LogicalBufOccupancies, LogicalBufFills>
-TemporalReuseAnalysis(const LogicalBufOccupancies& occupancies);
-
-struct SpatialReuseInfo
-{
-  LinkTransferInfo link_transfer_info;
-  MulticastInfo multicast_info;
-};
-
-SpatialReuseInfo SpatialReuseAnalysis(LogicalBufFills&,
-                                      LogicalBufOccupancies&,
-                                      const LinkTransferModel&,
-                                      const MulticastModel&);
 
 isl::map MapToShifted(isl::space space, size_t pos, int shift);
 
 isl_val* ValOfConstantPwPolynomial(isl_pw_qpolynomial* qp);
 
 unsigned long ValToUnsignedLong(isl_val* val);
-
-std::pair<Occupancy, Fill> FillFromOccupancy(Occupancy occupancy)
-{
-  /**
-   * Compute fill by iteratively going through temporal loops and
-   * computing delta, only stopping when delta is non-zero at time 1.
-   */
-
-  bool try_again = true;
-  while (try_again)
-  {
-    try_again = false;
-    for (auto dim_it = occupancy.in_rbegin(); dim_it != occupancy.in_rend();
-        ++dim_it)
-    {
-      auto [dim_idx, dim_type] = *dim_it;
-      if (dim_type == spacetime::Dimension::Time)
-      {
-        std::cout << occupancy << std::endl;
-        auto time_shift_map = occupancy.tag_like_this(
-          isl::map_to_shifted(occupancy.space().domain(), dim_idx, -1)
-        );
-        auto occ_before = time_shift_map.apply_range(occupancy.map);
-        auto fill = occupancy.subtract(occ_before.map);
-        auto first_fill = isl::fix_si(fill.map, isl_dim_in, dim_idx, 1);
-        if (first_fill.range().is_empty())
-        {
-          occupancy.project_dim_in(dim_idx, 1);
-          try_again = true;
-          break;
-        }
-        else
-        {
-          return std::make_pair(occupancy, fill);
-        }
-      }
-    }
-  }
-
-  return std::make_pair(occupancy, occupancy);
-}
-
-std::pair<LogicalBufOccupancies, LogicalBufFills>
-TemporalReuseAnalysis(const LogicalBufOccupancies& occupancies)
-{
-  LogicalBufFills fills;
-  LogicalBufOccupancies effectual_occupancies;
-
-  for (auto& [buf, occupancy] : occupancies)
-  {
-    auto [eff_occupancy, fill] = FillFromOccupancy(occupancy);
-    fills.emplace(std::make_pair(buf, fill));
-    effectual_occupancies.emplace(std::make_pair(buf, eff_occupancy));
-  }
-
-  return std::make_pair(effectual_occupancies, fills);
-}
-
-SpatialReuseInfo
-SpatialReuseAnalysis(LogicalBufFills& fills,
-                     LogicalBufOccupancies& occupancies,
-                     const LinkTransferModel& link_transfer_model,
-                     const MulticastModel& multicast_model)
-{
-  auto link_transfer_info = link_transfer_model.Apply(fills, occupancies);
-  auto multicast_info = multicast_model.Apply(
-    link_transfer_info.unfulfilled_fills,
-    occupancies
-  );
-
-  return SpatialReuseInfo{
-    .link_transfer_info = std::move(link_transfer_info),
-    .multicast_info = std::move(multicast_info)
-  };
-}
 
 NestAnalysis::NestAnalysis()
 {
